backend/requirements.txt
Flask>=3.1.1
flask-cors>=4.0.0
cx_Oracle==8.3.0
ldap3==2.9.1
PyJWT>=2.3.0
pytest==7.4.0
pyotp==2.8.0
qrcode==7.4.2
Pillow>=10.0.0
python-dotenv>=1.0.0
